fix a flaw with the two sprite touching each other
Before this change, it only works “one direction”, meaning if you have sprite A being carried into a pipe with sprite B overlapping the pipes, it does not interact with the sprites as intended. However if the player is carrying sprite B into a pipe and touches sprite A, it WILL interact which is unintended. The game handles the sprite in a “1-way” fasion to save performance so that each pair of sprite indexes are checked only once and not twice.