update
I found out that to set the timer at the minimum if below it DOES NOT HAPPEN EVERY FRAME. When inside a pipe, codes at $0196CB ONLY executes if the timer hits $03, therefore the timer ends up with decrementing till $03, and then reset to [#($04+!Setting_Minimal_StuntimerSprites)] and repeats, which means that the amount of time after exiting the pipe and before the sprite unstuns itself is inconsistent. It is now fixed to work in the intended way: If the sprite timer goes below [#($04+!Setting_Minimal_StuntimerSprites)], it should set it to that minimum every frame throughout the pipe traveling.