update - potential bugs found and potential new feature added
During testing, the bug that caused the player to travel upwards if you set his pipe state to $09 IMMIDIATELY after the player enters a pipe cap (which can result in $09-$04=$05). Although this is fixed at line 237 (JMP .PipeCodeReturn), this is just in case of other, potential glitches.

Also if Mario's prep direction is 0, and gets transfered to his pipe state when Mario reaches his drag destination, now resets his state completely. This can be used in a way the player enters a pipe and magically appears out of thin air like you would with exit enabled pipes that on the level the player is taken to is set to “Do Nothing - Cannot Bring Item”