Fixed a minor bug that de-centers the player
Fixed a minor bug in which springboard and keys can still push the player to be off by 1 pixel when traveling in small horizontal pipes. This is noticeable if you have a turn corner (turn right-down or up-right) with a key/springboard placed above them when the player travels from horizontal to downwards.