update - fixed enterable range
When playing around with SMW's vertical exit-enabled pipes, Mario's position must in the range from $0003 to $001C inclusively (relative to the left-half cap; $0000 is centered within a tile on the left-half of the pipe cap). My pipe's enterable range is 1-pixel positioned off to the right.

I fixed this minor issue and also fixed the comment info in the defines that it is actually $0005, from the center, if you wanted the SSPs to mimic SMW's pipes on “if mario is centered enough to enter”.