fix
Made a better solution on handling how the player dies inside the pipe. Originally, if you die inside a pipe, Mario remains invisible (remains visible, but remains behind the layer) and looks somewhat weird. This update makes it so that if you die inside a pipe, will actually do his normal animation, including in front of the layer. This is done by checking if both states are true:
-Player is in pipe mode
-Player is dying ($71 == $09)
then it resets his pipe mode and executes the resetstatus code for only 1 frame.