fixed a potential glitch
While $13FB is USUALLY set along with $9D, this isn't always the case if using custom ASM that would freeze the player, such as my upcoming scroll limits patch. When $13FB is set, $00DC2D is not executed (this is the code that takes the values of $7B/$7D, and increments $94/$96), however, the timer "!Freeram_SSP_PipeTmr" will STILL decrements while the player is NOT moving out of the pipe. when $13FB is nonzero during the entering/exiting transition, which can cause the player to, upon setting $13FB back to zero, returns to his normal out-of-pipe state BEFORE he fully comes out of the pipe, causing clipping issues and can kill the player.